pyproject.toml
[project]
name = "stock-det"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cap-from-youtube>=0.2.0",
    "cvzone>=1.6.1",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.3",
    "ultralytics>=8.3.250",
]
